Criando saque.

+++ src/main/java/br/edu/unisep/bank/controller/AccountController.java
@@ -34,7 +34,7 @@ public class AccountController {
 
     @PostMapping("/conta")
     public Account createAccount(@Validated @RequestBody Account account){
-        account.setSaldo((float) 0.00);
+        account.setSaldo(0.00);
         return repository.save(account);
     }
